Use requesting client regarless of it's visibility
In the past we used try and use best possible client, however that was problematic, for example when different tab was focused no matching client could be discovered. Intent was to let requests go even if client is throttled, but it seems like a mistake as well. If client is throttled there is a good reason and it does not makes sense to workaround that it's still going to be throlled.